changed size_t to unsigned long to compile on Windows

+++ kratos/containers/distributed_sparse_graph.h
@@ -147,7 +147,7 @@ public:
     {
         IndexType MinJ, MaxJ;
         ComputeLocalMinMaxColumnIndex(MinJ,MaxJ);
-        return GetComm().MaxAll(MaxJ);
+        return GetComm().MaxAll(static_cast<unsigned long>(MaxJ));
     }
